+++ planeri-backend/database/migrations/2024_04_09_180540_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->require();
+            $table->unsignedBigInteger('user_id')->require();
             $table->float('price');
             $table->string('status');
             $table->timestamps();
